+++ API.Tests/ParserTest.cs
@@ -1,3 +1,4 @@
+using API.Parser;
 using Xunit;
 using static API.Parser.Parser;
 
@@ -15,8 +16,10 @@ namespace API.Tests
         //[InlineData("Dance in the Vampire Bund v16-17 (Digital) (NiceDragon)", "16-17")]
         [InlineData("Akame ga KILL! ZERO v01 (2016) (Digital) (LuCaZ).cbz", "1")]
         [InlineData("v001", "1")]
+        [InlineData("No Volume", "0")]
         [InlineData("U12 (Under 12) Vol. 0001 Ch. 0001 - Reiwa Scans (gb)", "1")]
         [InlineData("[Suihei Kiki]_Kasumi_Otoko_no_Ko_[Taruby]_v1.1.zip", "1")]
+        [InlineData("Tonikaku Cawaii [Volume 11].cbz", "11")]
         public void ParseVolumeTest(string filename, string expected)
         {
             Assert.Equal(expected, ParseVolume(filename));
@@ -33,11 +36,20 @@ namespace API.Tests
         [InlineData("v001", "")]
         [InlineData("U12 (Under 12) Vol. 0001 Ch. 0001 - Reiwa Scans (gb)", "U12 (Under 12)")]
         [InlineData("Akame ga KILL! ZERO (2016-2019) (Digital) (LuCaZ)", "Akame ga KILL! ZERO")]
+        [InlineData("APOSIMZ 017 (2018) (Digital) (danke-Empire).cbz", "APOSIMZ")]
+        [InlineData("Akiiro Bousou Biyori - 01.jpg", "Akiiro Bousou Biyori")]
+        [InlineData("Beelzebub_172_RHS.zip", "Beelzebub")]
+        [InlineData("Dr. STONE 136 (2020) (Digital) (LuCaZ).cbz", "Dr. STONE")]
+        [InlineData("Cynthia the Mission 29.rar", "Cynthia the Mission")]
+        [InlineData("Darling in the FranXX - Volume 01.cbz", "Darling in the FranXX")]
+        [InlineData("Darwin's Game - Volume 14 (F).cbz", "Darwin's Game")]
+        [InlineData("[BAA]_Darker_than_Black_c7.zip", "Darker than Black")]
+        [InlineData("Kedouin Makoto - Corpse Party Musume, Chapter 19 [Dametrans].zip", "Kedouin Makoto - Corpse Party Musume")]
         public void ParseSeriesTest(string filename, string expected)
         {
             Assert.Equal(expected, ParseSeries(filename));
         }
-        
+
         [Theory]
         [InlineData("Killing Bites Vol. 0001 Ch. 0001 - Galactica Scanlations (gb)", "1")]
         [InlineData("My Girlfriend Is Shobitch v01 - ch. 09 - pg. 008.png", "9")]
@@ -49,6 +61,8 @@ namespace API.Tests
         [InlineData("c001", "1")]
         [InlineData("[Suihei Kiki]_Kasumi_Otoko_no_Ko_[Taruby]_v1.12.zip", "12")]
         [InlineData("Adding volume 1 with File: Ana Satsujin Vol. 1 Ch. 5 - Manga Box (gb).cbz", "5")]
+        [InlineData("Hinowa ga CRUSH! 018 (2019) (Digital) (LuCaZ).cbz", "18")]
+        [InlineData("Cynthia The Mission - c000-006 (v06) [Desudesu&Brolen].zip", "0-6")]
         public void ParseChaptersTest(string filename, string expected)
         {
             Assert.Equal(expected, ParseChapter(filename));
@@ -85,10 +99,11 @@ namespace API.Tests
         
         [Theory]
         [InlineData("test.cbz", true)]
-        [InlineData("test.cbr", true)]
+        [InlineData("test.cbr", false)]
         [InlineData("test.zip", true)]
-        [InlineData("test.rar", true)]
+        [InlineData("test.rar", false)]
         [InlineData("test.rar.!qb", false)]
+        [InlineData("[shf-ma-khs-aqs]negi_pa_vol15007.jpg", false)]
         public void IsArchiveTest(string input, bool expected)
         {
             Assert.Equal(expected, IsArchive(input));
