+++ API/Controllers/AccountController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using API.Constants;
 using API.DTOs;
@@ -76,15 +77,14 @@ namespace API.Controllers
             if (registerDto.IsAdmin)
             {
                 _logger.LogInformation($"{user.UserName} is being registered as admin. Granting access to all libraries.");
-                var libraries = await _unitOfWork.LibraryRepository.GetLibrariesAsync();
+                var libraries = (await _unitOfWork.LibraryRepository.GetLibrariesAsync()).ToList();
                 foreach (var lib in libraries)
                 {
                     lib.AppUsers ??= new List<AppUser>();
                     lib.AppUsers.Add(user);
                 }
+                if (libraries.Any() && !await _unitOfWork.Complete()) _logger.LogInformation("There was an issue granting library access. Please do this manually.");
             }
-            
-            if (!await _unitOfWork.Complete()) _logger.LogInformation("There was an issue granting library access. Please do this manually.");
 
             return new UserDto
             {
@@ -97,7 +97,11 @@ namespace API.Controllers
         public async Task<ActionResult<UserDto>> Login(LoginDto loginDto)
         {
             var user = await _userManager.Users
-                .SingleOrDefaultAsync(x => x.UserName == loginDto.Username.ToLower());
+                .SingleOrDefaultAsync(x => x.NormalizedUserName == loginDto.Username.ToUpper());
+
+            var debugUsers = await _userManager.Users.Select(x => x.NormalizedUserName).ToListAsync();
+            
+            _logger.LogInformation($"All Users: {String.Join(",", debugUsers)}");
 
             if (user == null) return Unauthorized("Invalid username");
