+++ API/DTOs/RegisterDto.cs
@@ -7,7 +7,7 @@ namespace API.DTOs
         [Required]
         public string Username { get; set; }
         [Required]
-        [StringLength(8, MinimumLength = 4)]
+        [StringLength(16, MinimumLength = 4)]
         public string Password { get; set; }
         public bool IsAdmin { get; set; }
     }
