+++ API/Interfaces/ILibraryRepository.cs
@@ -14,7 +14,6 @@ namespace API.Interfaces
         Task<Library> GetLibraryForIdAsync(int libraryId);
         Task<IEnumerable<LibraryDto>> GetLibraryDtosForUsernameAsync(string userName);
         Task<IEnumerable<Library>> GetLibrariesAsync();
-        Task<Library> GetLibraryForNameAsync(string libraryName);
         Task<bool> DeleteLibrary(int libraryId);
     }
 }
