+++ API/Interfaces/ITaskScheduler.cs
@@ -2,12 +2,8 @@
 {
     public interface ITaskScheduler
     {
-        public void ScanLibrary(int libraryId, bool forceUpdate = false);
-
-        public void CleanupVolumes(int[] volumeIds);
-        /// <summary>
-        /// Clears the cache directory entirely.
-        /// </summary>
-        public void ClearCache();
+        void ScanLibrary(int libraryId, bool forceUpdate = false);
+        void CleanupVolumes(int[] volumeIds);
+        void ScanSeries(int libraryId, int seriesId);
     }
 }
