+++ API/appsettings.Development.json
@@ -9,6 +9,12 @@
       "Microsoft": "Information",
       "Microsoft.Hosting.Lifetime": "Information",
       "Hangfire": "Information"
+    },
+    "File": {
+      "Path": "kavita.log",
+      "Append": "True",
+      "FileSizeLimitBytes": 0,
+      "MaxRollingFiles": 0
     }
   }
 }
